requirements.txt
# HPC论文自动获取工具依赖

# arXiv API
feedparser>=6.0.10
requests>=2.31.0

# AI模型API (支持多种模型)
openai>=1.0.0  # 用于DeepSeek和Qwen (OpenAI兼容接口)
google-generativeai>=0.3.0  # 用于Gemini

# 邮件发送
# 使用标准库smtplib，无需额外依赖

# 定时任务
APScheduler>=3.10.4
pytz>=2023.3

# 数据库
# 使用标准库sqlite3，无需额外依赖
